+++ include/ylt/coro_rpc/impl/coro_rpc_client.hpp
@@ -233,11 +233,13 @@ class coro_rpc_client {
    * @param host server address
    * @param port server port
    * @param connect_timeout_duration RPC call timeout seconds
+   * @param eps server endpoints
    * @return error code
    */
   [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect(
       std::string host, std::string port,
-      std::chrono::steady_clock::duration connect_timeout_duration) {
+      std::chrono::steady_clock::duration connect_timeout_duration,
+      std::vector<asio::ip::tcp::endpoint> *eps = nullptr) {
     auto lock_ok = connect_mutex_.tryLock();
     if (!lock_ok) {
       co_await connect_mutex_.coScopedLock();
@@ -253,29 +255,40 @@ class coro_rpc_client {
     }
     auto ret = co_await connect_impl(
         std::chrono::duration_cast<std::chrono::milliseconds>(
-            connect_timeout_duration));
+            connect_timeout_duration),
+        eps);
     connect_mutex_.unlock();
     co_return std::move(ret);
   }
   [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect(
-      std::string_view endpoint,
+      std::string_view address,
       std::chrono::steady_clock::duration connect_timeout_duration) {
-    auto pos = endpoint.find(':');
-    std::string host(endpoint.substr(0, pos));
-    std::string port(endpoint.substr(pos + 1));
+    auto pos = address.find(':');
+    std::string host(address.substr(0, pos));
+    std::string port(address.substr(pos + 1));
 
     return connect(std::move(host), std::move(port), connect_timeout_duration);
   }
 
   [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect(
-      std::string_view endpoint) {
-    return connect(endpoint, config_.connect_timeout_duration);
+      std::string_view address) {
+    return connect(address, config_.connect_timeout_duration);
   }
 
   [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect(
       std::string host, std::string port) {
     return connect(std::move(host), std::move(port),
                    config_.connect_timeout_duration);
+  }
+
+  [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect(
+      std::string_view address, std::vector<asio::ip::tcp::endpoint> *eps) {
+    auto pos = address.find(':');
+    std::string host(address.substr(0, pos));
+    std::string port(address.substr(pos + 1));
+
+    return connect(std::move(host), std::move(port),
+                   config_.connect_timeout_duration, eps);
   }
 
 #ifdef YLT_ENABLE_SSL
@@ -393,7 +406,8 @@ class coro_rpc_client {
   static bool is_ok(coro_rpc::err_code ec) noexcept { return !ec; }
 
   [[nodiscard]] async_simple::coro::Lazy<coro_rpc::err_code> connect_impl(
-      std::chrono::milliseconds conn_timeout_dur) {
+      std::chrono::milliseconds conn_timeout_dur,
+      std::vector<asio::ip::tcp::endpoint> *eps) {
     if (should_reset_) {
       co_await reset();
     }
@@ -415,12 +429,36 @@ class coro_rpc_client {
           .start([](auto &&) {
           });
     }
-
-    std::error_code ec = co_await coro_io::async_connect(
-        &control_->executor_, control_->socket_, config_.host, config_.port);
+    std::vector<asio::ip::tcp::endpoint> eps_tmp;
+    if (eps == nullptr) {
+      eps = &eps_tmp;
+    }
+    std::error_code ec;
+    asio::ip::tcp::resolver::iterator iter;
+    if (eps->empty()) {
+      ELOG_TRACE << "start resolve host: " << config_.host << ":"
+                 << config_.port;
+      std::tie(ec, iter) = co_await coro_io::async_resolve(
+          &control_->executor_, control_->socket_, config_.host, config_.port);
+      asio::ip::tcp::resolver::iterator end;
+      while (iter != end) {
+        eps->push_back(iter->endpoint());
+        ++iter;
+      }
+      if (eps->empty()) {
+        co_return errc::not_connected;
+      }
+    }
+    ELOG_TRACE << "start connect to endpoint lists. total endpoint count:"
+               << eps->size() << ", the first endpoint is: "
+               << (*eps)[0].address().to_string();
+    if (auto [ec, _] = co_await coro_io::async_connect(&control_->executor_,
+                                                       control_->socket_, *eps);
+        ec) {
+      co_return errc::not_connected;
+    }
     std::error_code err_code;
     timer_->cancel(err_code);
-
     if (control_->is_timeout_) {
       ELOG_WARN << "client_id " << config_.client_id << " connect timeout";
       co_return errc::timed_out;
@@ -430,6 +468,7 @@ class coro_rpc_client {
                 << " failed:" << ec.message();
       co_return errc::not_connected;
     }
+
     if (config_.enable_tcp_no_delay == true) {
       control_->socket_.set_option(asio::ip::tcp::no_delay(true), ec);
     }
