+++ src/coro_io/tests/test_client_pool.cpp
@@ -28,9 +28,11 @@
 #include "async_simple/coro/Semaphore.h"
 #include "async_simple/coro/Sleep.h"
 #include "async_simple/coro/SpinLock.h"
+#include "ylt/coro_http/coro_http_client.hpp"
 #include "ylt/coro_rpc/impl/coro_rpc_client.hpp"
 #include "ylt/coro_rpc/impl/default_config/coro_rpc_config.hpp"
 #include "ylt/coro_rpc/impl/expected.hpp"
+#include "ylt/struct_pack.hpp"
 using namespace std::chrono_literals;
 using namespace async_simple::coro;
 template <typename T = coro_rpc::coro_rpc_client>
@@ -160,8 +162,8 @@ TEST_CASE("test reconnect") {
 struct mock_client : public coro_rpc::coro_rpc_client {
   using coro_rpc::coro_rpc_client::coro_rpc_client;
   async_simple::coro::Lazy<coro_rpc::errc> connect(
-      const std::string &hostname) {
-    auto ec = co_await this->coro_rpc::coro_rpc_client::connect(hostname);
+      const std::string &hostname, std::vector<asio::ip::tcp::endpoint> *eps) {
+    auto ec = co_await this->coro_rpc::coro_rpc_client::connect(hostname, eps);
     if (ec) {
       co_await coro_io::sleep_for(300ms);
     }
@@ -241,4 +243,116 @@ TEST_CASE("test client pools parallel r/w") {
     co_await collectAll(std::move(works));
   }());
   ELOG_DEBUG << "test client pools parallel r/w over.";
+}
+
+TEST_CASE("test client pools dns cache") {
+  async_simple::coro::syncAwait([]() -> async_simple::coro::Lazy<void> {
+    auto pool = coro_io::client_pool<coro_http::coro_http_client>::create(
+        "http://www.baidu.com",
+        coro_io::client_pool<coro_http::coro_http_client>::pool_config{
+            .dns_cache_update_duration = 600s});
+    auto eps_init = pool->get_remote_endpoints();
+    CHECK(eps_init->empty());
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          cli.close();
+          co_return;
+        });
+    auto eps = pool->get_remote_endpoints();
+    CHECK(!eps->empty());
+    CHECK(eps.get() != eps_init.get());
+    CHECK(eps->front().port() == 80);
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          cli.close();
+          co_return;
+        });
+    auto eps2 = pool->get_remote_endpoints();
+    CHECK(eps.get() == eps2.get());
+  }());
+}
+
+TEST_CASE("test client pools dns refresh") {
+  async_simple::coro::syncAwait([]() -> async_simple::coro::Lazy<void> {
+    auto pool = coro_io::client_pool<coro_http::coro_http_client>::create(
+        "http://www.baidu.com",
+        coro_io::client_pool<coro_http::coro_http_client>::pool_config{
+            .dns_cache_update_duration = 0s});
+    auto eps_init = pool->get_remote_endpoints();
+    CHECK(eps_init->empty());
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          cli.close();
+          co_return;
+        });
+    auto eps = pool->get_remote_endpoints();
+    CHECK(!eps->empty());
+    CHECK(eps.get() != eps_init.get());
+    CHECK(eps->front().port() == 80);
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          co_return;
+        });
+    auto eps2 = pool->get_remote_endpoints();
+    CHECK(eps.get() != eps2.get());
+  }());
+}
+
+TEST_CASE("test client pools dns parallel refresh") {
+  async_simple::coro::syncAwait([]() -> async_simple::coro::Lazy<void> {
+    auto pool = coro_io::client_pool<coro_http::coro_http_client>::create(
+        "http://www.baidu.com",
+        coro_io::client_pool<coro_http::coro_http_client>::pool_config{
+            .dns_cache_update_duration = 0s});
+    auto eps_init = pool->get_remote_endpoints();
+    CHECK(eps_init->empty());
+    std::vector<
+        async_simple::coro::RescheduleLazy<ylt::expected<void, std::errc>>>
+        results;
+    for (int i = 0; i < 200; ++i) {
+      results.push_back(
+          pool->send_request(
+                  [](coro_http::coro_http_client &cli) -> Lazy<void> {
+                    auto result = co_await cli.async_get("/");
+                    CHECK(result.net_err == std::error_code{});
+                    co_return;
+                  })
+              .via(coro_io::get_global_executor()));
+    }
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          auto result = co_await cli.async_get("/");
+          CHECK(result.net_err == std::error_code{});
+          co_return;
+        });
+    auto eps = pool->get_remote_endpoints();
+    CHECK(!eps->empty());
+    CHECK(eps.get() != eps_init.get());
+    CHECK(eps->front().port() == 80);
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          co_return;
+        });
+    auto eps2 = pool->get_remote_endpoints();
+    CHECK(eps.get() != eps2.get());
+  }());
+}
+
+TEST_CASE("test client pools dns don't refresh") {
+  async_simple::coro::syncAwait([]() -> async_simple::coro::Lazy<void> {
+    auto pool = coro_io::client_pool<coro_http::coro_http_client>::create(
+        "http://www.baidu.com",
+        coro_io::client_pool<coro_http::coro_http_client>::pool_config{
+            .dns_cache_update_duration = -1s});
+    auto eps_init = pool->get_remote_endpoints();
+    CHECK(eps_init->empty());
+    co_await pool->send_request(
+        [](coro_http::coro_http_client &cli) -> Lazy<void> {
+          cli.close();
+          co_return;
+        });
+    auto eps = pool->get_remote_endpoints();
+    CHECK(eps.get() == eps_init.get());
+    CHECK(eps->empty());
+  }());
 }
